New structure for the blockchain as a tree, BlockchainTree.
Thre tree is composed of a TreeStruct used for sending through the network.
The struct (with method) can be recreated by passing the struct to the constructor of the BlockchainTree.

The TreeStruct is backed by 3 structures: 2 array for the blocks and tree node and a map of uuid key to position in the arrays.
	blockchain: The actual blockchain
	nodeList: array of tree node (at same position as associated blockchain block. This is used to old the parent and child for traversal
	uuidMap: Mapping of uuid to position in the 2 arrays for quick reference during server calls

More test is needed but the basic is there.